More refactoring of the communication service.
Refactored to have most of the common zmodem handling stuff in an abstract class.
There have been two implementations, a file-based one and a jssc serial library
based version.  The jssc version is yet to be added, but will effectively have
the pipes built in the original version of the SerialCommunicationService class
(now called AbstractZmodemService).