Windows now works.
File uploads are not working when the lsz command is used directly.  It creates
an empty file on the target and then hangs.  Using the shell script as the IDE
did does work, so on windows there is no file upload progress (though it seems
faster than the other platforms anyway).