Added windows native application launcher
The windows application launcher must be built on windows.  It was added
as a profile that gets activated when a build is done on a windows machine.
The jar file from the open source project at
https://java.net/projects/appbundler/pages/Home is now installed as an artifact
in the build repository as it is declared as a dependency by the plugin
japp-maven-plugin which is used to build the mac and windows native launcher.